+++ serialize/.env
@@ -1,0 +1,4 @@
+export WIKIREAD_HOST=cutewiki.charlie.city
+export WIKIREAD_PORT=8822
+export WIKIREAD_REQ_NAME=permission_data_pull
+export WIKIREAD_REQ_PASSCODE=615f1fb2-b1ad-43c7-a764-9be5ed80f8b1
