+++ utils/helper.h
@@ -1,0 +1,12 @@
+#ifndef __HELPER_L__
+#define __HELPER_L__
+
+/* RESIZE FUNCTIONALITY */
+void *resize_array(void *arr, int *max_len, int curr_index, size_t singleton_size);
+
+/* SPLIT STRING FUNCTIONALITY */
+int char_is_range(char _char);
+int num_is_range(char _char);
+char **split_string(char *full_string, char delimeter, int *arr_len, char *extra, ...);
+
+#endif /* __HELPER_L__ */
